Use Gradle TestKit for integration tests / remove output polling (#262)
This change uses Gradle TestKit to speed up the Gradle integration tests and for Maven removes the output polling.

Note that the tests that called run have been changed to call test and a test is included when you create a project. The tests that called run were failing intermittently anyway since they ran on port 8080 which was frequently causing port binding failures on Github Actions.