Use Maven with JUnit default tests names
Right now there is a problem with the Maven pom.xml generated by the starter when you use JUnit. The problem is that Surefire (the Maven plugin that runs unit tests) is configured to include tests in any file `<include>%regex[.*]</include>`. **This cause that, in unit testing phase, Surefire tries to run also integration tests, and build fails.**

In Maven the expected default is that Surefire runs tests in files like:
  - **/Test*.java
  - **/*Test.java
  - **/*Tests.java
  - **/*TestCase.java
and Failsafe plugin runs integration tests in files like:
  - **/IT*.java
  - **/*IT.java
  - **/*ITCase.java

With this commit the Maven pom.xml is generated with the espected defaults and is only reconfigured if Spock or Kotest are used. And in wich case the espected Spock or Kotest default filenames are used.