Add the micronaut-data-processor to the Maven annotation processors
When building a MongoDB Micronaut Data application, if the micronaut-data-processor is not
specified, then the Repository introspection is missing the Mongo Queries.

This is fixed by adding the transient "micronaut-data-processor" in to the list of
annotation processors, but ONLY if it is added before the micronaut-data-document-processor.

If it is added after, we seem to get a NPE from the data-processor. I am going to raise an
issue for this in the micronaut-data project.