Kapt dependencies in Maven seem to require an order
A method was renamed in https://github.com/micronaut-projects/micronaut-core/pull/7051

But as some annotation libs pull in older versions of inject-java, Maven fails if they are
earlier in the dependency list.

This was found investigating here https://github.com/micronaut-projects/micronaut-guides/issues/918

I suspect it's also causing the other Kotlin failures in guides.

This also reverts some of the changes here https://github.com/micronaut-projects/micronaut-starter/pull/1146
when we had the same issue with Mongo-Data processing in Maven

I need to test building the guides against a snapshot of this.