Fix the mongo spec.
We missed parenthesis on the call to get the size of a List.

On the 3.4.x branch, this is ok, and the size method is called (I'm guessing due to some Spock
magic -- but I cannot prove it currently)

On the 3.5.x branch, this weird behavior no longer works, so we just get a list of 3 empty elements.

Strangely, only on Java 17

This fix is applied to the 3.4.x branch (as it's wrong there, and just works due to some edge case
and merging it up to 3.5.x will fix the test over there)