Use url-connection-client for dynamodb and graal (#1319)
* Use url-connection-client for dynamodb and graal

The apache client doesn't work with native compilation, so if the user selects
dynamodb and graalvm then we switch to the url-connection one, and exclude the other
from the dependencies

This also fixes exclusions for Gradle which were not previously required

* Address feedback

* Use isNotEmpty

* Move logic to utils (#1320)

* refactor: move logic to utils

This moves the logic to a Utils class so that we are able to apply this to other AWS SDK related features not just to dynamodb. E.g. S3.

* use CollectionsUtils

* Fix my bad merge

Co-authored-by: Tim Yates <tim.yates@gmail.com>

Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>