bug: Use correct serialization annotation in default/function
If a Micronaut Serialization feature was selected, DEFAULT FUNCTION applications still used `@Introspected` in the generated Controller.

This PR switches to `@Serdeable` if the feature has been added

Whilst writing the tests for this, I discovered that the GCP module does not support micronaut-serialization

Raised https://github.com/micronaut-projects/micronaut-gcp/issues/724 to look into this, and added a feature constraint to prevent it.