Fix TestResources with Hibernate Reactive
If we are using hibernate-reactive and Test Resources, we need to specify the correct module test resources to use.
We also need to provide the JDBC driver for the chosen database as a `testResourcesService` dependency.

In Maven, this is done by adding both to the `testResourcesDependencies` block of the `micronaut-maven-plugin` block

Whilst doing this, I also discovered we were using the wrong `db-type` for postgres (we had `postgresql`, but it should be `postgres`)

We have an issue with using the vert.x drives with postgres, as can be seen in the feature tests (I commented postgres out), I will raise a separate issue for this.