Remove aws-java-sdk-core as a test dependency
When we add aws-java-sdk-core as a test dependency, Maven removes it from the compile time transitive dependencies.

This causes the jms-aws-sqs maven guide to fail as it cannot find com.amazonaws.regions.Regions at compile time.

As the dependency comes in transitively, we can remove it from the locastack feature, as it should be coming in transitively from whichever micronaut aws library the user is using