Stop adding microstream annotations as compile as well as provided
For Groovy Maven builds, annotation processors end up as being provided dependencies.
However, we add microstream-annotation as a compile dependency.

For Groovy compilation to succeed, we need to only have the provided dependency.

This was discovered as part of getting the MicroStream guide working for 4.0.0

See: https://github.com/micronaut-projects/micronaut-guides/pull/1269\#issuecomment-1620402381