Stop hardcoding a version of micronaut-security-aot
We were hardcoding a version of micronaut-security-aot, which meant that we ended up
with two versions of a core library when trying to run the AOT analysis.

This led to multiple beans being in the context as seen in

https://github.com/micronaut-projects/micronaut-security/issues/1420

The fix here imports the micronaut-platform for the aotPlugin configuration

I don't believe we need this for the maven dependency.