Exclude dependencies in Maven for metrics annotation
The metrics-annotation has an api dependency on aop, which is pulling in other versions of things.

This PR excludes micronaut-core from the dependencies for the metrics annotation processor.

Excluding inject as with the other processors is not enough as we also need to exclude io.micronaut.core.beans.BeanIntrospection$Builder