Add gradle-kotlin to tests and fix Coherence
This commit adds gradle-kotlin tests to several in starter-core that were just testing the Groovy DSL.

It also fixes the Coherence feature where we were adding a version as a property.
This is problematic in the gradle-kotlin dsl as we would need to define it in the build script.
However I think in these days of Coherence having a BOM, it's also incorrect, and the version should come from there, not starter