Fix gradle configuration for Kotlin with Java 21
There is also a task kotlinTestResourcesCompile which is sometimes added to Gradle builds.

Under Java 21, this fails as compilerOptions.jvmTarget is not configured for it.

This change switches to configuring ALL KotlinCompile by type, instead of just named tasks.

It also fixes Kapt for Java 21 by switching to use the java 17 toolchain for kotlin.