Fix Kotlin DSL when specifying docker image names
Previously we had 2 issues.

In some places we were using `$project.name` which resolves to `${project}.name` in Kotlin (so is incorrect)

In others, we were using `images = [ ... ]` which is not valid kotlin.

Changed the build rocker, and fixed and extended the tests to check both Kts and groovy DSL.

Closes #2231