breaking(github): remove 'Push to GitHub' project generation
This requires you grant write access to our application for the sake of 3 minutes of convenience.

This is not something we should be advocating from a security point of view.

The PR to remove it from the UI can be found here https://github.com/micronaut-projects/micronaut-starter-ui/pull/108

That should be deployed before this is released.