fix: mn create fails since 4.3.0
In 4.3.0 when we added chatbot creation, we abstracted the builder commands away.

At one point, we combine application-type features and user specified features into a single set.

The issue here is that the set in the case of the create command was unmodifiable.

This then failed with the esoteric 'Error null' message.

This PR takes the features from the application type, and loads them into a HashSet prior to doing any manipulation.

I did it this way as it's hard to test this area of the functionality, and it will prevent us having a similar issue if we add further commands

Closes #2318